+++ zcash_client_backend/src/data_api/wallet.rs
@@ -667,11 +667,11 @@ where
     let mut sapling_output_meta = vec![];
     let mut transparent_output_meta = vec![];
     for payment in proposal.transaction_request().payments() {
-        let recipient_address = payment
-            .recipient_address
+        let recipient_zaddr = payment.recipient_address.clone();
+        match recipient_zaddr
             .clone()
-            .convert_if_network::<Address>(params.network_type())?;
-        match recipient_address {
+            .convert_if_network::<Address>(params.network_type())?
+        {
             Address::Unified(ua) => {
                 let memo = payment
                     .memo
@@ -686,10 +686,7 @@ where
                         memo.clone(),
                     )?;
                     sapling_output_meta.push((
-                        Recipient::Unified(
-                            ua.clone(),
-                            PoolType::Shielded(ShieldedProtocol::Sapling),
-                        ),
+                        Recipient::External(recipient_zaddr, PoolType::SAPLING),
                         payment.amount,
                         Some(memo),
                     ));
@@ -711,7 +708,11 @@ where
                     .as_ref()
                     .map_or_else(MemoBytes::empty, |m| m.clone());
                 builder.add_sapling_output(external_ovk, addr, payment.amount, memo.clone())?;
-                sapling_output_meta.push((Recipient::Sapling(addr), payment.amount, Some(memo)));
+                sapling_output_meta.push((
+                    Recipient::External(recipient_zaddr, PoolType::SAPLING),
+                    payment.amount,
+                    Some(memo),
+                ));
             }
             Address::Transparent(to) => {
                 if payment.memo.is_some() {
@@ -719,7 +720,11 @@ where
                 } else {
                     builder.add_transparent_output(&to, payment.amount)?;
                 }
-                transparent_output_meta.push((to, payment.amount));
+                transparent_output_meta.push((
+                    Recipient::External(recipient_zaddr, PoolType::TRANSPARENT),
+                    to,
+                    payment.amount,
+                ));
             }
         }
     }
@@ -737,19 +742,14 @@ where
                     memo.clone(),
                 )?;
                 sapling_output_meta.push((
-                    Recipient::InternalAccount(
-                        account,
-                        PoolType::Shielded(ShieldedProtocol::Sapling),
-                    ),
+                    Recipient::Internal(account, PoolType::SAPLING),
                     change_value.value(),
                     Some(memo),
                 ))
             }
             ShieldedProtocol::Orchard => {
                 #[cfg(not(feature = "orchard"))]
-                return Err(Error::UnsupportedPoolType(PoolType::Shielded(
-                    ShieldedProtocol::Orchard,
-                )));
+                return Err(Error::UnsupportedPoolType(PoolType::ORCHARD));
 
                 #[cfg(feature = "orchard")]
                 unimplemented!("FIXME: implement Orchard change output creation.")
@@ -771,10 +771,7 @@ where
                     .output_index(i)
                     .expect("An output should exist in the transaction for each Sapling payment.");
 
-                let received_as = if let Recipient::InternalAccount(
-                    account,
-                    PoolType::Shielded(ShieldedProtocol::Sapling),
-                ) = recipient
+                let received_as = if let Recipient::Internal(account, PoolType::SAPLING) = recipient
                 {
                     build_result
                         .transaction()
@@ -794,28 +791,27 @@ where
                 SentTransactionOutput::from_parts(output_index, recipient, value, memo, received_as)
             });
 
-    let transparent_outputs = transparent_output_meta.into_iter().map(|(addr, value)| {
-        let script = addr.script();
-        let output_index = build_result
-            .transaction()
-            .transparent_bundle()
-            .and_then(|b| {
-                b.vout
-                    .iter()
-                    .enumerate()
-                    .find(|(_, tx_out)| tx_out.script_pubkey == script)
-            })
-            .map(|(index, _)| index)
-            .expect("An output should exist in the transaction for each transparent payment.");
+    let transparent_outputs =
+        transparent_output_meta
+            .into_iter()
+            .map(|(recipient, addr, value)| {
+                let script = addr.script();
+                let output_index = build_result
+                    .transaction()
+                    .transparent_bundle()
+                    .and_then(|b| {
+                        b.vout
+                            .iter()
+                            .enumerate()
+                            .find(|(_, tx_out)| tx_out.script_pubkey == script)
+                    })
+                    .map(|(index, _)| index)
+                    .expect(
+                        "An output should exist in the transaction for each transparent payment.",
+                    );
 
-        SentTransactionOutput::from_parts(
-            output_index,
-            Recipient::Transparent(addr),
-            value,
-            None,
-            None,
-        )
-    });
+                SentTransactionOutput::from_parts(output_index, recipient, value, None, None)
+            });
 
     wallet_db
         .store_sent_tx(&SentTransaction {