+++ zcash_keys/CHANGELOG.md
@@ -21,6 +21,7 @@ The entries below are relative to the `zcash_client_backend` crate as of
 - A new `orchard` feature flag has been added to make it possible to
   build client code without `orchard` dependendencies.
 - `zcash_keys::address::Address::to_zcash_address`:
+- `zcash_keys::address::Address::try_from_zcash_address`:
 
 ### Changed
 - `zcash_keys::address`: